requirements.txt
Click==7.0
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pymongo==3.9.0
python-dotenv==0.10.3
Werkzeug==0.16.0
